feat(nifs): impr protogalaxy poly err handling
